Update zfs operator to make it compatible with openzfs 2.0
Remove zfs-localpv images from docker images to be added to k3s package

We don't use zfs-localpv for charts we support. Users looking to use it can pull it from network automatically when they use it. This should reduce image size by a considerable size, almost ~200MB.

Fix zfs operator

Update generated bindata

Remove zfs-operator from manifests

K3s has an issue where it does not apply all CRDs specified. So we will be applying the zfs-operator manifest ourselves in this case to ensure all necessary bits are in place.

Update pkg bind data

Add zstd as build dep

Remove docker.io from image tag names

Fix zfs-localpv crd (#11)

We had a bug in k3s where it was not applying the manifests in manifests directory properly ( similar to how we would expect with kubectl ) and the fault lied with it not specifically supporting beta version of CRD api which zfs-localpv specified upstream. However upstream zfs-localpv has fixed it by using the stable api verison ( the issue remains in k3s still but as we are not consuming the beta api anymore, we are not going to be hit by it ).