Fix zfs-localpv crd (#11)
We had a bug in k3s where it was not applying the manifests in manifests directory properly ( similar to how we would expect with kubectl ) and the fault lied with it not specifically supporting beta version of CRD api which zfs-localpv specified upstream. However upstream zfs-localpv has fixed it by using the stable api verison ( the issue remains in k3s still but as we are not consuming the beta api anymore, we are not going to be hit by it ).